XIVY-14720 Support switch theme in dev-wf-ui
